moving r by c models to separate file
